# release-23.2: changefeedccl: reduce use of errors during pauseOrResumePolling