--- a/pkg/ccl/changefeedccl/changefeedbase/target.go
+++ b/pkg/ccl/changefeedccl/changefeedbase/target.go
@@ -110,6 +110,19 @@ func (ts *Targets) EachTableID(f func(descpb.ID) error) error {
 	return nil
 }
 
+// EachTableIDWithBool is similar to EachTableID but avoids using
+// iterutil.Map(err) to eliminate the overhead of errors.Is. Thus, f should not
+// return iterutil.StopIteration(). It returns false with error when the
+// callback f returns false or true when the iteration completes.
+func (ts *Targets) EachTableIDWithBool(f func(descpb.ID) (bool, error)) (bool, error) {
+	for id := range ts.m {
+		if b, err := f(id); !b {
+			return false, err
+		}
+	}
+	return true, nil
+}
+
 // EachHavingTableID iterates over each Target with the given id, returning
 // false if there were none.
 func (ts *Targets) EachHavingTableID(id descpb.ID, f func(Target) error) (bool, error) {