Backport 2/2 commits from #131951.

/cc @cockroachdb/release
 
---

**schemafeed: add BenchmarkPauseOrResumePolling**
This patch introduces a benchmark for pauseOrResumePolling, as it’s an important
function in a hot path that has previously caused performance regressions due to
complex error handling.

Epic: none
Release note: none

---

**changefeedccl: reduce use of errors during pauseOrResumePolling**

Previously, errors were constructed and compared in pauseOrResumePolling solely
to indicate that pausing cannot be stopped and period table history scan is
necessary during pauseOrResumePolling. However, no actual errors were returned
from pauseOrResumePolling. This error handling introduced unnecessary CPU
overhead in this hot path, as observed during escalations and DRT scale testing.
This patch improves performance by replacing error-based communication with a
boolean check.

Fixes: https://github.com/cockroachdb/cockroach/issues/131327
Release note (performance improvement): Enhanced performance when schema_locked
is not in use by improving error handling during periodic table history polling.

```
                                        │  before.txt  │              after.txt               │
                                        │    sec/op    │    sec/op     vs base                │
PauseOrResumePolling/non-terminal_error    5.533µ ± 5%   3.849µ ± 29%  -30.44% (p=0.000 n=10)
PauseOrResumePolling/not_schema_locked    7069.5n ± 4%   176.4n ±  9%  -97.50% (p=0.000 n=10)
PauseOrResumePolling/schema_locked         136.2n ± 4%   131.8n ±  7%        ~ (p=0.306 n=10)
```

---

Release justification: low risk performance improvement